requirements.txt
asgiref==3.8.1
Brotli==1.1.0
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.3.2
cssselect2==0.7.0
Django==5.1.1
django-renderpdf==4.0.0
django-unfold==0.40.0
fonttools==4.54.1
html5lib==1.1
idna==3.10
mysql-connector==2.2.9
pillow==10.4.0
pycparser==2.22
pydyf==0.11.0
pyphen==0.16.0
requests==2.32.3
six==1.16.0
sqlparse==0.5.1
tinycss2==1.3.0
urllib3==2.2.3
weasyprint==62.3
webencodings==0.5.1
zopfli==0.2.3
